+++ igittigitt/__init__.py
@@ -19,3 +19,5 @@ __name__ = __init__conf__.name
 __url__ = __init__conf__.url
 __author__ = __init__conf__.author
 __author_email__ = __init__conf__.author_email
+
+__all__ = ["conf_igittigitt", "IgnoreParser"]