+++ igittigitt/igittigitt.py
@@ -17,7 +17,6 @@ try:
 except ImportError:  # pragma: no cover
     from conf_igittigitt import conf_igittigitt  # type: ignore  # pragma: no cover
 
-
 PathLikeOrString = Union[str, "os.PathLike[Any]"]
 __all__ = ("IgnoreParser",)
 
@@ -126,6 +125,14 @@ class IgnoreParser(object):
     ) -> None:
         pass
 
+    @staticmethod
+    def _expand_base_path(base_path: PathLikeOrString) -> pathlib.Path:
+        """
+        expand the user directory and make absolute, but dont resolve symlinks
+        """
+        path_base_dir = pathlib.Path(os.path.abspath(os.path.expanduser(base_path)))
+        return path_base_dir
+
     # parse_rule_files{{{
     def parse_rule_files(
         self, base_dir: PathLikeOrString, filename: str = ".gitignore", add_default_patterns: bool = conf_igittigitt.add_default_patterns
@@ -173,7 +180,7 @@ class IgnoreParser(object):
         """
         # parse_rule_files}}}
 
-        path_base_dir = pathlib.Path(base_dir).resolve()
+        path_base_dir = self._expand_base_path(base_path=base_dir)
 
         if add_default_patterns:
             self._add_default_patterns(path_base_dir=path_base_dir)
@@ -204,12 +211,12 @@ class IgnoreParser(object):
             see README.RST, Section "Default Patterns"
 
         """
-        path_rule_file = pathlib.Path(rule_file).resolve()
+        path_rule_file = self._expand_base_path(base_path=rule_file)
 
         if not base_dir:
             path_base_dir = path_rule_file.parent
         else:
-            path_base_dir = pathlib.Path(base_dir).resolve()
+            path_base_dir = self._expand_base_path(base_path=base_dir)
 
         with open(path_rule_file) as ignore_file:
             counter = 0
@@ -244,9 +251,7 @@ class IgnoreParser(object):
             directory, that needs to be provided here
         """
         # add_rule}}}
-
-        path_base_dir = pathlib.Path(base_path).resolve()
-
+        path_base_dir = self._expand_base_path(base_path=base_path)
         rules = get_rules_from_git_pattern(git_pattern=pattern, path_base_dir=path_base_dir)
 
         if rules:
@@ -259,10 +264,23 @@ class IgnoreParser(object):
     def match(self, file_path: PathLikeOrString) -> bool:
         """
         returns True if the path matches the rules
+
+        >>> # Setup
+        >>> base_path = pathlib.Path(__file__).parent.parent.resolve() / 'tests/example_negation'
+
+        >>> # Test
+        >>> gitignore = IgnoreParser()
+        >>> gitignore.add_rule("/*", base_path)
+        >>> gitignore.add_rule("!/foo", base_path)
+        >>> gitignore.add_rule("/foo/*", base_path)
+        >>> gitignore.add_rule("!/foo/bar", base_path)
+        >>> assert gitignore.match(base_path / "foo/bar/file.txt") == False
+        >>> # assert gitignore.match(base_path / "foo/other/file.txt") == True  # this fails - because everything is wrong
+        >>> # see : https://docs.rs/ignore/0.4.18/ignore/struct.WalkBuilder.html
+
         """
         # match}}}
-
-        path_file_object = pathlib.Path(os.path.abspath(os.path.expanduser(file_path)))
+        path_file_object = self._expand_base_path(base_path=file_path)
         is_file = path_file_object.is_file()
         str_file_path = str(path_file_object)
 
@@ -276,12 +294,13 @@ class IgnoreParser(object):
 
     def _match_rules(self, str_file_path: str, is_file: bool) -> bool:
         """
-        match without negotiations - in that case we can return
+        match without negations - in that case we can return
         immediately after a match.
 
 
         is_file:
             the passed path is a file (and not a directory)
+
         """
 
         # small optimisation - we have a good chance
@@ -406,8 +425,7 @@ class IgnoreParser(object):
         Ignore function for shutil.copy_tree
         """
         # shutil_ignore}}}
-
-        path_base_dir = pathlib.Path(base_dir).resolve()
+        path_base_dir = self._expand_base_path(base_path=base_dir)
         ignore_files: Set[str] = set()
         for file in file_names:
             if self.match(path_base_dir / file):
@@ -536,13 +554,13 @@ def git_pattern_handle_blanks(git_pattern: str) -> str:
     see: https://stackoverflow.com/questions/10213653
     wcmatch.glob.globmatch supports both forms
 
-    >>> assert git_pattern_handle_blanks(r'something \\ \\ ') == 'something\\ \\ '
-    >>> assert git_pattern_handle_blanks(r'something \\ \\  ') == 'something\\ \\ '
-    >>> assert git_pattern_handle_blanks(r'some\\ thing \\ ') == 'some\\ thing\\ '
-    >>> assert git_pattern_handle_blanks(r'some thing \\ ') == 'some thing\\ '
+    >>> assert git_pattern_handle_blanks(r'something \\ \\ ') == r'something\\ \\ '
+    >>> assert git_pattern_handle_blanks(r'something \\ \\  ') == r'something\\ \\ '
+    >>> assert git_pattern_handle_blanks(r'some\\ thing \\ ') == r'some\\ thing\\ '
+    >>> assert git_pattern_handle_blanks(r'some thing \\ ') == r'some thing\\ '
     """
-    parts = [part.strip() for part in git_pattern.split("\\ ")]
-    return "\\ ".join(parts)
+    parts = [part.strip() for part in git_pattern.split(r"\ ")]
+    return r"\ ".join(parts)
 
 
 def get_match_anchored(git_pattern: str) -> bool: